Some policies

+++ routes/web.php
@@ -56,25 +56,25 @@ Route::controller(RegisterController::class)->group(function () {
 
 Route::controller(ProjectController::class)->group(function () {
     
-    Route::get('/create-project', 'showCreateProjectForm')->name('project.create');
-    Route::post('/projects', 'store')->name('project.store');
-    Route::get('/home', 'home')->name('project.home');
-    Route::get('/project/{title}', 'show')->name('project.show');
-    Route::get('/myprojects', 'index')->name('project.index');
-    Route::get('/project/{title}/addMember', 'showaddMemberForm')->name('project.addMember');
-    Route::post('/project/{title}/addMember/store', 'addOneMember')->name('project.Memberstore');
-    Route::get('/project/{title}/addLeader', 'showaddLeaderForm')->name('project.addLeader');
-    Route::post('/project/{title}/addLeader/store', 'addOneLeader')->name('project.Leaderstore');
-    Route::get('/pending-invites', 'pendingInvite')->name('pending.invites');
-    Route::post('/accept-invite/{id_user}/{id_project}', 'acceptInvite')->name('accept.invite');
-    Route::post('/decline-invite/{id_user}/{id_project}', 'declineInvite')->name('decline.invite');
-    Route::get('/projects/{title}/members', 'showMembers')->name("project.showMember");
-    Route::get('/projects/{title}/leaders', 'showLeaders')->name("project.showLeader");
-    Route::get('/project/{title}/edit', 'edit')->name('project.editProject');
-   Route::put('/project/{title}/update', 'update')->name('project.update');
-   Route::get('/project/{title}/favorite', 'favorite')->name('project.favorite');
-   Route::get('/project/{title}/noFavorite', 'noFavorite')->name('project.noFavorite');
-   Route::post('project/{title}/archived','archived')->name('project.archived');
+    Route::get('/create-project', 'showCreateProjectForm')->name('project.create')->middleware('auth');
+    Route::post('/projects', 'store')->name('project.store')->middleware('auth');
+    Route::get('/home', 'home')->name('project.home')->middleware('auth');
+    Route::get('/project/{title}', 'show')->name('project.show')->middleware('auth');
+    Route::get('/myprojects', 'index')->name('project.index')->middleware('auth');
+    Route::get('/project/{title}/addMember', 'showaddMemberForm')->name('project.addMember')->middleware('auth');
+    Route::post('/project/{title}/addMember/store', 'addOneMember')->name('project.Memberstore')->middleware('auth');
+    Route::get('/project/{title}/addLeader', 'showaddLeaderForm')->name('project.addLeader')->middleware('auth');
+    Route::post('/project/{title}/addLeader/store', 'addOneLeader')->name('project.Leaderstore')->middleware('auth');
+    Route::get('/pending-invites', 'pendingInvite')->name('pending.invites')->middleware('auth');
+    Route::post('/accept-invite/{id_user}/{id_project}', 'acceptInvite')->name('accept.invite')->middleware('auth');
+    Route::post('/decline-invite/{id_user}/{id_project}', 'declineInvite')->name('decline.invite')->middleware('auth');
+    Route::get('/projects/{title}/members', 'showMembers')->name("project.showMember")->middleware('auth');
+    Route::get('/projects/{title}/leaders', 'showLeaders')->name("project.showLeader")->middleware('auth');
+    Route::get('/project/{title}/edit', 'edit')->name('project.editProject')->middleware('auth');
+   Route::put('/project/{title}/update', 'update')->name('project.update')->middleware('auth');
+   Route::get('/project/{title}/favorite', 'favorite')->name('project.favorite')->middleware('auth');
+   Route::get('/project/{title}/noFavorite', 'noFavorite')->name('project.noFavorite')->middleware('auth');
+   Route::post('project/{title}/archived','archived')->name('project.archived')->middleware('auth');
    
 
 
@@ -85,18 +85,18 @@ Route::controller(ProjectController::class)->group(function () {
 });
 
 Route::controller(UsersController::class)->group(function () {
-Route::get('/search/users', 'search')->name('search.users');
+Route::get('/search/users', 'search')->name('search.users')->middleware('auth');
 
 
 });
 
 Route::controller(TaskController::class)->group(function () {
     
-     Route::get('/project/{title}/createTask', 'create')->name('task.create');
-    Route::post('/project/{title}/storeTask', 'store')->name('task.store');
-    Route::get('/project/{title}/task', 'show')->name('task.show');
-    Route::patch('/project/{title}/task/{taskId}/complete', 'isCompleted')->name('task.complete');
-    Route::delete('/project/{title}/task/{taskId}/delete', 'delete')->name('task.delete');
+     Route::get('/project/{title}/createTask', 'create')->name('task.create')->middleware('auth');
+    Route::post('/project/{title}/storeTask', 'store')->name('task.store')->middleware('auth');
+    Route::get('/project/{title}/task', 'show')->name('task.show')->middleware('auth');
+    Route::patch('/project/{title}/task/{taskId}/complete', 'isCompleted')->name('task.complete')->middleware('auth');
+    Route::delete('/project/{title}/task/{taskId}/delete', 'delete')->name('task.delete')->middleware('auth');
 
 
 
@@ -109,23 +109,23 @@ Route::controller(AdminController::class)->group(function () {
 
 
 Route::controller(ProfileController::class)->group(function(){
-    Route::get('/profile/{id}', 'show')->name('profile');
-    Route::get('/profile/{id}/edit', 'edit')->name('profile.edit');
-   Route::put('/profile/{id}/update', 'update')->name('profile.update');
-   Route::delete('/profile/{id}/delete', 'delete')->name('profile.delete');
+    Route::get('/profile/{id}', 'show')->name('profile')->middleware('auth');
+    Route::get('/profile/{id}/edit', 'edit')->name('profile.edit')->middleware('auth');
+   Route::put('/profile/{id}/update', 'update')->name('profile.update')->middleware('auth');
+   Route::delete('/profile/{id}/delete', 'delete')->name('profile.delete')->middleware('auth');
   
 
 });
 
 Route::controller(CommentController::class)->group(function(){
-    Route::get('/project/{title}/task/{taskId}/comments','show')->name('task.comment');
-    Route::post('/project/{title}/task/{taskId}/comments/store', [CommentController::class, 'store'])->name('comments.store');
+    Route::get('/project/{title}/task/{taskId}/comments','show')->name('task.comment')->middleware('auth');
+    Route::post('/project/{title}/task/{taskId}/comments/store', [CommentController::class, 'store'])->name('comments.store')->middleware('auth');
 
 });
 
 Route::controller(LikeController::class)->group(function(){
-Route::post('/api/like/store', 'store')->name('like.store');
-Route::delete('/api/like/{id}', 'destroy')->name('like.destroy');
+Route::post('/api/like/store', 'store')->name('like.store')->middleware('auth');
+Route::delete('/api/like/{id}', 'destroy')->name('like.destroy')->middleware('auth');
 
 });
